break up state.rs
